requirements.txt
# ExamHandOCR - Requirements
# Python 3.9+

# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Data processing
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
imgaug>=0.4.0

# Model dependencies
timm>=0.9.0
transformers>=4.35.0
einops>=0.7.0

# Training & Evaluation
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scipy>=1.10.0

# OCR & Text Processing
pytesseract>=0.3.10
regex>=2023.8.8
editdistance>=0.6.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & Utilities
pyyaml>=6.0.1
omegaconf>=2.3.0
hydra-core>=1.3.0
jsonlines>=3.1.0
pandas>=2.0.0

# Optional: Layout models
# detectron2 (install manually)
# mmdet (install manually)

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
